+++ packages/js/src/modules/modal.ts
@@ -11,12 +11,12 @@ export class Modal {
   }
 
   open() {
-    toggleClass(this.root, 's-open', true);
+    toggleClass(this.root, 'vxs-open', true);
     setAriaHidden(this.root, false);
   }
 
   close() {
-    toggleClass(this.root, 's-open', false);
+    toggleClass(this.root, 'vxs-open', false);
     setAriaHidden(this.root, true);
   }
 
@@ -42,3 +42,4 @@ export class Modal {
     });
   }
 }
+